+++ libs/components/ag-grid/src/lib/modules/ag-grid/ag-grid.module.ts
@@ -2,6 +2,7 @@ import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { SkyViewkeeperModule } from '@skyux/core';
 import { SkyDataManagerModule } from '@skyux/data-manager';
+import { SkyDatePipeModule } from '@skyux/datetime';
 import { SkyIconModule } from '@skyux/indicators';
 import { SkyInlineDeleteModule } from '@skyux/layout';
 import { SkyThemeModule } from '@skyux/theme';
@@ -14,7 +15,6 @@ import { SkyAgGridDataManagerAdapterDirective } from './ag-grid-data-manager-ada
 import { SkyAgGridRowDeleteComponent } from './ag-grid-row-delete.component';
 import { SkyAgGridRowDeleteDirective } from './ag-grid-row-delete.directive';
 import { SkyAgGridWrapperComponent } from './ag-grid-wrapper.component';
-import { SkyAgGridService } from './ag-grid.service';
 import { SkyAgGridCellEditorAutocompleteModule } from './cell-editors/cell-editor-autocomplete/cell-editor-autocomplete.module';
 import { SkyAgGridCellEditorCurrencyModule } from './cell-editors/cell-editor-currency/cell-editor-currency.module';
 import { SkyAgGridCellEditorDatepickerModule } from './cell-editors/cell-editor-datepicker/cell-editor-datepicker.module';
@@ -54,6 +54,7 @@ import { SkyAgGridHeaderComponent } from './header/header.component';
     SkyAgGridCellEditorTextModule,
     SkyAgGridResourcesModule,
     SkyDataManagerModule,
+    SkyDatePipeModule,
     SkyIconModule,
     SkyInlineDeleteModule,
     SkyViewkeeperModule,
@@ -67,6 +68,5 @@ import { SkyAgGridHeaderComponent } from './header/header.component';
     SkyAgGridHeaderComponent,
     SkyAgGridHeaderGroupComponent,
   ],
-  providers: [SkyAgGridService],
 })
 export class SkyAgGridModule {}