+++ libs/components/core/src/lib/modules/numeric/numeric.module.ts
@@ -3,11 +3,10 @@ import { NgModule } from '@angular/core';
 import { SkyCoreResourcesModule } from '../shared/sky-core-resources.module';
 
 import { SkyNumericPipe } from './numeric.pipe';
-import { SkyNumericService } from './numeric.service';
 
 @NgModule({
   declarations: [SkyNumericPipe],
-  providers: [SkyNumericPipe, SkyNumericService],
+  providers: [SkyNumericPipe],
   imports: [SkyCoreResourcesModule],
   exports: [SkyNumericPipe],
 })