+++ libs/components/modals/src/lib/modules/modal/modal.service.ts
@@ -1,9 +1,4 @@
-import {
-  ComponentRef,
-  EnvironmentInjector,
-  Injectable,
-  inject,
-} from '@angular/core';
+import { ComponentRef, Injectable } from '@angular/core';
 import { SkyDynamicComponentService } from '@skyux/core';
 
 import { SkyModalHostContext } from './modal-host-context';
@@ -25,7 +20,6 @@ export class SkyModalService {
   private static host: ComponentRef<SkyModalHostComponent> | undefined;
 
   #dynamicComponentService: SkyDynamicComponentService;
-  #environmentInjector = inject(EnvironmentInjector);
 
   constructor(dynamicComponentService: SkyDynamicComponentService) {
     this.#dynamicComponentService = dynamicComponentService;
@@ -102,7 +96,6 @@ export class SkyModalService {
       SkyModalService.host = this.#dynamicComponentService.createComponent(
         SkyModalHostComponent,
         {
-          environmentInjector: this.#environmentInjector,
           providers: [
             {
               provide: SkyModalHostContext,