+++ libs/sdk/tools/jest.config.ts
@@ -1,0 +1,20 @@
+/* eslint-disable */
+export default {
+  displayName: 'tools',
+  preset: '../../../jest.preset.js',
+  transform: {
+    '^.+\\.[tj]s$': ['ts-jest', { tsconfig: '<rootDir>/tsconfig.spec.json' }],
+  },
+  moduleFileExtensions: ['ts', 'js', 'html'],
+  testEnvironment: 'node',
+  collectCoverage: true,
+  coverageDirectory: '../../../coverage/libs/sdk/tools',
+  coverageThreshold: {
+    global: {
+      branches: 100,
+      functions: 100,
+      lines: 100,
+      statements: 100,
+    },
+  },
+};